requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
boto3==1.40.60
botocore==1.40.60
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
dataclasses-json==0.6.7
distro==1.9.0
duckduckgo_search==8.1.1
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.28.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-modelarmor==0.3.0
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==1.0.2
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.1
langchain-google-community==3.0.0
langchain-openai==1.0.1
langchain-text-splitters==1.0.0
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.38
lxml==6.0.2
Markdown==3.9
MarkupSafe==3.0.3
marshmallow==3.26.1
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.10.0
numpy==2.3.4
openai==2.6.1
orjson==3.11.4
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
primp==0.15.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.28.0
rsa==4.9.1
s3transfer==0.14.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
streamlit==1.50.0
tenacity==9.1.2
tiktoken==0.12.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
